temporary backup

+++ src/pages/Menu/Diarys/reducer/diarysSlice.js
@@ -4,7 +4,9 @@ import { v4 } from "uuid";
 
 const initialState = {
     diarys: [],
-    loading: true,
+    loading: false,
+    isEnd: false,
+    page: 0,
 };
 
 const NumToEmotion = {
@@ -16,7 +18,10 @@ const NumToEmotion = {
 
 export const fetchDiaryGet = createAsyncThunk(
     "diarys/fetchDiaryGet",
-    async (page = 0) => {
+    async (page = 0, { rejectWithValue }) => {
+        if (!axios.defaults.headers.common.Authorization)
+            return rejectWithValue("Authorization Error");
+        console.log(`page: ${page}`);
         const res = await axios.get(`/diary?page=${page}`);
         return res.data;
     }
@@ -24,11 +29,13 @@ export const fetchDiaryGet = createAsyncThunk(
 
 export const fetchDiaryAdd = createAsyncThunk(
     "diarys/fetchDiaryAdd",
-    async (body, dispatch) => {
+    async (body, { rejectWithValue }) => {
+        if (!axios.defaults.headers.common.Authorization)
+            return rejectWithValue("Authorization Error");
         const { postDate, emotion } = body;
-        const offset = postDate.getTimezoneOffset() * 60000;
-        const today = new Date(postDate - offset);
-        body.postDate = today.toISOString();
+        // console.log(postDate);
+        delete body.id;
+        body.postDate = postDate.slice(0, 19);
         body.emotion = NumToEmotion[emotion];
         console.log(body);
         const res = await axios.post("/diary", body);
@@ -37,12 +44,15 @@ export const fetchDiaryAdd = createAsyncThunk(
 );
 export const fetchDiaryUpdate = createAsyncThunk(
     "diarys/fetchDiaryUpdate",
-    async (body, dispatch) => {
-        const { postDate, emotion } = body;
-        const offset = postDate.getTimezoneOffset() * 60000;
-        const today = new Date(postDate - offset);
-        body.postDate = today.toISOString();
+    async (body, { rejectWithValue }) => {
+        if (!axios.defaults.headers.common.Authorization)
+            return rejectWithValue("Authorization Error");
+        const { id, postDate, emotion } = body;
+        body.postId = id;
+        const today = new Date(postDate.getTime() + 1000 * 60 * 60 * 9);
+        body.postDate = today.toISOString().slice(0, 19);
         body.emotion = NumToEmotion[emotion];
+        console.log(body);
         const res = await axios.put("/diary", body);
         return res.data;
     }
@@ -50,8 +60,9 @@ export const fetchDiaryUpdate = createAsyncThunk(
 
 export const fetchDiaryDelete = createAsyncThunk(
     "diarys/fetchDiaryDelete",
-    async (postId, dispatch) => {
-        console.log(postId);
+    async (postId, { rejectWithValue }) => {
+        if (!axios.defaults.headers.common.Authorization)
+            return rejectWithValue("Authorization Error");
         const res = await axios.delete(`/diary/${postId}`);
         return res.data;
     }
@@ -61,6 +72,14 @@ const diarysSlice = createSlice({
     name: "diarys",
     initialState,
     reducers: {
+        diaryClear: (state, action) => {
+            state.diarys = [];
+            state.page = 0;
+            return state;
+        },
+        pageInc: (state, action) => {
+            state.page += 1;
+        },
         diaryAdd: (state, action) => {
             const diary = action.payload;
             state.diarys.push({
@@ -87,13 +106,20 @@ const diarysSlice = createSlice({
         });
         builder.addCase(fetchDiaryGet.fulfilled, (state, action) => {
             state.loading = false;
-            state.diarys = action.payload.data
-                .slice()
-                .sort((a, b) => new Date(b.postDate) - new Date(a.postDate));
+            if (action.payload.data.length) {
+                const newDiaryCollection = action.payload.data
+                    .slice()
+                    .sort(
+                        (a, b) => new Date(b.postDate) - new Date(a.postDate)
+                    );
+                state.diarys.push(...newDiaryCollection);
+            } else {
+                state.isEnd = true;
+            }
             state.error = "";
         });
         builder.addCase(fetchDiaryGet.rejected, (state, action) => {
-            state.loading = false;
+            if (action.payload !== "Authorization Error") state.loading = false;
             state.error = action.error;
         });
         builder.addCase(fetchDiaryAdd.pending, (state) => {
@@ -101,22 +127,26 @@ const diarysSlice = createSlice({
         });
         builder.addCase(fetchDiaryAdd.fulfilled, (state, action) => {
             state.loading = false;
+            state.page = 0;
+            state.diarys = [];
             state.error = "";
         });
         builder.addCase(fetchDiaryAdd.rejected, (state, action) => {
-            state.loading = false;
+            if (action.payload !== "Authorization Error") state.loading = false;
             state.error = action.error;
         });
         builder.addCase(fetchDiaryUpdate.pending, (state, action) => {
             state.loading = true;
         });
         builder.addCase(fetchDiaryUpdate.fulfilled, (state, action) => {
+            console.log(action);
             state.loading = false;
-            console.log(action.payload);
+            state.page = 0;
+            state.diarys = [];
             state.error = "";
         });
         builder.addCase(fetchDiaryUpdate.rejected, (state, action) => {
-            state.loading = false;
+            if (action.payload !== "Authorization Error") state.loading = false;
             state.error = action.error;
         });
         builder.addCase(fetchDiaryDelete.pending, (state) => {
@@ -124,15 +154,16 @@ const diarysSlice = createSlice({
         });
         builder.addCase(fetchDiaryDelete.fulfilled, (state, action) => {
             state.loading = false;
-            console.log(action.payload.status);
+            state.page = 0;
+            state.diarys = [];
             state.error = "";
         });
         builder.addCase(fetchDiaryDelete.rejected, (state, action) => {
-            state.loading = false;
+            if (action.payload !== "Authorization Error") state.loading = false;
             state.error = action.error;
         });
     },
 });
 
 export default diarysSlice.reducer;
-export const { diaryAdd, diaryUpdate, diaryDelete } = diarysSlice.actions;
+export const { diaryClear, pageInc } = diarysSlice.actions;
